ThreeScaleToolbox::CRD module contains all custom resource dump and parsing logic
