+++ lib/3scale_toolbox.rb
@@ -20,6 +20,7 @@ require '3scale_toolbox/resource_reader'
 require '3scale_toolbox/configuration'
 require '3scale_toolbox/remotes'
 require '3scale_toolbox/3scale_client_factory'
+require '3scale_toolbox/crds'
 require '3scale_toolbox/entities'
 require '3scale_toolbox/attribute_filters'
 require '3scale_toolbox/base_command'
