+++ lib/3scale_toolbox/crds.rb
@@ -1,0 +1,16 @@
+require '3scale_toolbox/crds/application_plan_dump'
+require '3scale_toolbox/crds/backend_method_dump'
+require '3scale_toolbox/crds/backend_mapping_rule_dump'
+require '3scale_toolbox/crds/backend_metric_dump'
+require '3scale_toolbox/crds/backend_dump'
+require '3scale_toolbox/crds/backend_usage_dump'
+require '3scale_toolbox/crds/limit_dump'
+require '3scale_toolbox/crds/mapping_rule_dump'
+require '3scale_toolbox/crds/method_dump'
+require '3scale_toolbox/crds/metric_dump'
+require '3scale_toolbox/crds/pricing_rule_dump'
+require '3scale_toolbox/crds/product_dump'
+require '3scale_toolbox/crds/product_deployment_crd_parser'
+require '3scale_toolbox/crds/backend_parser'
+require '3scale_toolbox/crds/product_parser'
+require '3scale_toolbox/crds/remote'
