+++ lib/3scale_toolbox/crds/application_plan_dump.rb
@@ -1,0 +1,17 @@
+module ThreeScaleToolbox
+  module CRD
+    module ApplicationPlan
+      def to_cr
+        {
+          'name' => name,
+          'appsRequireApproval' => approval_required?,
+          'trialPeriod' => trial_period_days,
+          'setupFee' => setup_fee,
+          'costMonth' => cost_per_month,
+          'pricingRules' => pricing_rules.map(&:to_cr),
+          'limits' => limits.map(&:to_cr)
+        }
+      end
+    end
+  end
+end
