+++ lib/3scale_toolbox/crds/backend_method_dump.rb
@@ -1,0 +1,12 @@
+module ThreeScaleToolbox
+  module CRD
+    module BackendMethod
+      def to_cr
+        {
+          'friendlyName' => friendly_name,
+          'description' => description,
+        }
+      end
+    end
+  end
+end
