+++ lib/3scale_toolbox/crds/backend_metric_dump.rb
@@ -1,0 +1,13 @@
+module ThreeScaleToolbox
+  module CRD
+    module BackendMetric
+      def to_cr
+        {
+          'friendlyName' => friendly_name,
+          'unit' => unit,
+          'description' => description,
+        }
+      end
+    end
+  end
+end
