+++ lib/3scale_toolbox/crds/backend_parser.rb
@@ -1,0 +1,4 @@
+module ThreeScaleToolbox
+  module CRD
+  end
+end
