+++ lib/3scale_toolbox/crds/backend_usage_dump.rb
@@ -1,0 +1,11 @@
+module ThreeScaleToolbox
+  module CRD
+    module BackendUsage
+      def to_cr
+        {
+          'path' => path
+        }
+      end
+    end
+  end
+end
