+++ lib/3scale_toolbox/crds/method_dump.rb
@@ -1,0 +1,12 @@
+module ThreeScaleToolbox
+  module CRD
+    module Method
+      def to_cr
+        {
+          'friendlyName' => friendly_name,
+          'description' => description,
+        }
+      end
+    end
+  end
+end
