+++ lib/3scale_toolbox/crds/metric_dump.rb
@@ -1,0 +1,13 @@
+module ThreeScaleToolbox
+  module CRD
+    module Metric
+      def to_cr
+        {
+          'friendlyName' => friendly_name,
+          'unit' => unit,
+          'description' => description,
+        }
+      end
+    end
+  end
+end
