+++ lib/3scale_toolbox/entities.rb
@@ -1,4 +1,3 @@
-require '3scale_toolbox/entities/product_deployment_crd_parser'
 require '3scale_toolbox/entities/service'
 require '3scale_toolbox/entities/application_plan'
 require '3scale_toolbox/entities/metric'
