+++ lib/3scale_toolbox/entities/backend_mapping_rule.rb
@@ -1,6 +1,8 @@
 module ThreeScaleToolbox
   module Entities
     class BackendMappingRule
+      include CRD::BackendMappingRule
+
       VALID_PARAMS = %w[metric_id pattern http_method delta position last].freeze
       public_constant :VALID_PARAMS
 
@@ -94,16 +96,6 @@ module ThreeScaleToolbox
         remote.delete_backend_mapping_rule backend.id, id
       end
 
-      def to_crd
-        {
-          'httpMethod' => http_method,
-          'pattern' => pattern,
-          'metricMethodRef' => metric_method_ref,
-          'increment' => delta,
-          'last' => last,
-        }
-      end
-
       private
 
       def mapping_rule_attrs
@@ -115,17 +107,6 @@ module ThreeScaleToolbox
         end
 
         mapping_rule
-      end
-
-      def metric_method_ref
-        if (method = backend.methods.find { |m| m.id == metric_id })
-          method.system_name
-        elsif (metric = backend.metrics.find { |m| m.id == metric_id })
-          metric.system_name
-        else
-          raise ThreeScaleToolbox::Error, "Unexpected error. Backend #{backend.system_name} " \
-            "mapping rule #{id} referencing to metric id #{metric_id} which has not been found"
-        end
       end
     end
   end
