+++ lib/3scale_toolbox/entities/pricing_rule.rb
@@ -1,6 +1,8 @@
 module ThreeScaleToolbox
   module Entities
     class PricingRule
+      include CRD::PricingRule
+
       class << self
         def create(plan:, metric_id:, attrs:)
           resp_attrs = plan.remote.create_pricingrule plan.id, metric_id, attrs
@@ -46,41 +48,15 @@ module ThreeScaleToolbox
         remote.delete_application_plan_pricingrule plan.id, metric_id, id
       end
 
-      def to_crd
-        {
-          'from' => min,
-          'to' => max,
-          'pricePerUnit' => cost_per_unit,
-          'metricMethodRef' => metric_method_ref,
-        }
-      end
-
       private
 
-      def metric_method_ref
-        # Find in service methods
-        # Find in service metrics
-        # Parse backend_id from pricing rule links
-        # Find in backend methods
-        # Find in backend metrics
-        if (method = plan.service.methods.find { |m| m.id == metric_id })
-          { 'systemName' => method.system_name }
-        elsif (metric = plan.service.metrics.find { |m| m.id == metric_id })
-          { 'systemName' => metric.system_name }
-        elsif (backend_id = Helper.backend_metric_link_parser(metric_link['href'] || ''))
-          backend = Backend.new(id: backend_id, remote: remote)
-          if (backend_metric = backend.metrics.find { |m| m.id == metric_id })
-            { 'systemName' => backend_metric.system_name, 'backend' => backend.system_name }
-          elsif (backend_method = backend.methods.find { |m| m.id == metric_id })
-            { 'systemName' => backend_method.system_name, 'backend' => backend.system_name }
-          else
-            raise ThreeScaleToolbox::Error, "Unexpected error. PricingRule #{id} " \
-              "referencing to metric id #{metric_id} which has not been found"
-          end
-        else
-          raise ThreeScaleToolbox::Error, "Unexpected error. PricingRule #{id} " \
-            "referencing to metric id #{metric_id} which has not been found"
-        end
+      # Used by CRD::PricingRule
+      # Returns the backend hosting the metric
+      def backend_from_metric
+        backend_id = Helper.backend_metric_link_parser(metric_link['href'] || '')
+        return if backend_id.nil?
+
+        Backend.new(id: backend_id, remote: remote)
       end
     end
   end
